pyproject.toml
[tool.poetry]
name = "gcn-transformer-network"
version = "0.1.0"
description = ""
authors = ["pytergeist <tompopeworks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tensorflow-macos = "^2.16.2"
tensorflow-metal = "^1.1.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
